requirements.txt
aiogram==3.26.0
aiosqlite==0.21.0
APScheduler==3.11.0
python-dotenv==1.1.1
